+++ nixos/modules/programs/default.nix
@@ -1,6 +1,5 @@
 {
   imports = [
-    ./dua.nix
     ./nix.nix
     ./ranger.nix
     ./tmux.nix